# Conflicts:
#	.gitignore
#	README.md
#	api/act.go
#	api/internal/handler/routes.go
#	api/internal/logic/act/completetasklogic.go
#	api/internal/logic/act/startprocinstlogic.go
#	api/test/a_test.go
#	common/act/client.go
#	common/act/config.go
#	common/act/ent.go
#	common/act/execution.go
#	common/act/execution/execution.go
#	common/act/execution/where.go
#	common/act/execution_create.go
#	common/act/execution_delete.go
#	common/act/execution_query.go
#	common/act/execution_update.go
#	common/act/hook/hook.go
#	common/act/identitylink.go
#	common/act/identitylink/identitylink.go
#	common/act/identitylink/where.go
#	common/act/identitylink_create.go
#	common/act/identitylink_delete.go
#	common/act/identitylink_query.go
#	common/act/identitylink_update.go
#	common/act/migrate/schema.go
#	common/act/mutation.go
#	common/act/predicate/predicate.go
#	common/act/procdef.go
#	common/act/procdef/procdef.go
#	common/act/procdef/where.go
#	common/act/procdef_create.go
#	common/act/procdef_delete.go
#	common/act/procdef_query.go
#	common/act/procdef_update.go
#	common/act/procinst.go
#	common/act/procinst/procinst.go
#	common/act/procinst/where.go
#	common/act/procinst_create.go
#	common/act/procinst_delete.go
#	common/act/procinst_query.go
#	common/act/procinst_update.go
#	common/act/runtime.go
#	common/act/schema/execution.go
#	common/act/schema/identitylink.go
#	common/act/schema/procdef.go
#	common/act/schema/procinst.go
#	common/act/schema/task.go
#	common/act/task.go
#	common/act/task/task.go
#	common/act/task/where.go
#	common/act/task_create.go
#	common/act/task_delete.go
#	common/act/task_query.go
#	common/act/task_update.go
#	common/act/tx.go
#	common/test/a_test.go
#	go.mod
#	go.sum
#	rpc/act.proto
#	rpc/etc/act.yaml
#	rpc/general/general.go
#	rpc/internal/logic/delprocdeflogic.go
#	rpc/internal/logic/delprocinstlogic.go
#	rpc/internal/logic/finddefbyformidlogic.go
#	rpc/internal/logic/findlatesttasklogic.go
#	rpc/internal/logic/saveprocdeflogic.go
#	rpc/internal/logic/savetasklogic.go
#	rpc/internal/logic/updateprocdeflogic.go
#	rpc/internal/logic/updateprocinstlogic.go
#	rpc/internal/logic/updatetasklogic.go
#	rpc/internal/logic/withdrawlogic.go
#	rpc/types/act/act.pb.go

+++ common/act/execution.go
@@ -25,7 +25,7 @@ type Execution struct {
 	// 开始时间
 	StartTime time.Time `json:"start_time,omitempty"`
 	// 是否删除,0:未删除,1:已删除
-	IsDel int8 `json:"is_del,omitempty"`
+	IsDel int32 `json:"is_del,omitempty"`
 	// 创建时间
 	CreateTime time.Time `json:"create_time,omitempty"`
 	// 流程修改时间
@@ -63,7 +63,7 @@ func (e *Execution) assignValues(columns []string, values []interface{}) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			e.ID = value.Int64
+			e.ID = int64(value.Int64)
 		case execution.FieldProcInstID:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field proc_inst_id", values[i])
@@ -92,7 +92,7 @@ func (e *Execution) assignValues(columns []string, values []interface{}) error {
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field is_del", values[i])
 			} else if value.Valid {
-				e.IsDel = int8(value.Int64)
+				e.IsDel = int32(value.Int64)
 			}
 		case execution.FieldCreateTime:
 			if value, ok := values[i].(*sql.NullTime); !ok {
